Chapter 3
In this chapter discussed SpriteKit’s physics engine and collision detection. Firstly, discussion of SKPhysicsBody. Then, turned on the game world’s gravity, which pulled the player down the length of the scene. After that, added a touch responder to propel the playerNode back up into space. Finally, in the end of the chapter had a discussion how to handle collisions.